refactor(auth): enhance OAuth provider integration and improve type definitions
