+++ packages/studiocms/src/routes/api/auth/[provider]/index.ts
@@ -5,19 +5,29 @@ import { Auth0OAuthAPI } from './_effects/auth0.js';
 import { DiscordOAuthAPI } from './_effects/discord.js';
 import { GitHubOAuthAPI } from './_effects/github.js';
 import { GoogleOAuthAPI } from './_effects/google.js';
-import { Provider, ProviderResponse, authEnvChecker } from './_shared.js';
+import { AuthAPIEffectDeps, Provider, ProviderResponse, authEnvChecker } from './_shared.js';
 
-const deps = Layer.mergeAll(
-	GoogleOAuthAPI.Default,
-	GitHubOAuthAPI.Default,
-	DiscordOAuthAPI.Default,
-	Auth0OAuthAPI.Default
-);
-
+/**
+ * Handles the GET request for OAuth authentication initialization for various providers.
+ *
+ * This route dynamically selects the appropriate OAuth provider (Google, GitHub, Discord, Auth0)
+ * based on the `provider` parameter in the request context. It checks if the provider is enabled
+ * in the authentication environment configuration before proceeding. If the provider is not enabled
+ * or not implemented, it returns a 501 response with an appropriate message.
+ *
+ * @param context - The API request context containing route parameters and other request data.
+ * @returns An API response from the selected provider's `initSession` method, or an error response if the provider is not configured or not implemented.
+ */
 export const GET: APIRoute = async (context: APIContext) =>
 	await convertToVanilla(
 		genLogger('studiocms/routes/api/auth/[provider]/index.GET')(function* () {
-			const authEnv = yield* authEnvChecker();
+			const [authEnv, google, github, discord, auth0] = yield* Effect.all([
+				authEnvChecker(),
+				GoogleOAuthAPI,
+				GitHubOAuthAPI,
+				DiscordOAuthAPI,
+				Auth0OAuthAPI,
+			]);
 
 			// Call the appropriate provider function based on the provider and function
 			switch (context.params.provider) {
@@ -25,34 +35,30 @@ export const GET: APIRoute = async (context: APIContext) =>
 					if (!authEnv.GOOGLE.ENABLED)
 						return ProviderResponse('Google provider is not configured', 501);
 
-					const google = yield* GoogleOAuthAPI;
 					return yield* google.initSession(context);
 				}
 				case Provider.GITHUB: {
 					if (!authEnv.GITHUB.ENABLED)
 						return ProviderResponse('GitHub provider is not configured', 501);
 
-					const github = yield* GitHubOAuthAPI;
 					return yield* github.initSession(context);
 				}
 				case Provider.DISCORD: {
 					if (!authEnv.DISCORD.ENABLED)
 						return ProviderResponse('Discord provider is not configured', 501);
 
-					const discord = yield* DiscordOAuthAPI;
 					return yield* discord.initSession(context);
 				}
 				case Provider.AUTH0: {
 					if (!authEnv.AUTH0.ENABLED)
 						return ProviderResponse('Auth0 provider is not configured', 501);
 
-					const auth0 = yield* Auth0OAuthAPI;
 					return yield* auth0.initSession(context);
 				}
 				default:
 					return ProviderResponse('Provider not implemented', 501);
 			}
-		}).pipe(Effect.provide(deps))
+		}).pipe(AuthAPIEffectDeps)
 	);
 
 export const OPTIONS: APIRoute = async () => OptionsResponse(['GET']);
